refactor(frontend): extract sidebar navigation into reusable modules
