+++ frontend/src/lib/extra-nav.ts
@@ -1,0 +1,5 @@
+import type { NavigationGroup } from "@/lib/sidebar-navigation"
+
+const extraNav: NavigationGroup[] = []
+
+export default extraNav
